% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ballpark_last_week}
\alias{ballpark_last_week}
\title{Compare this week's total with last week's}
\usage{
ballpark_last_week(df)
}
\arguments{
\item{df}{dataframe output of \link[ddvR]{ddv_transform}}
}
\value{
single value expressed as percentage of previous week's total
}
\description{
This is to check the ballpark of the imported values is approximately correct
should not be used for anything real. If first week of the year, the comparison
is with wek 52 of the previous year, regardless of whether week 53 exists. Because
it doesn't matter and i can't be bothered.
}
