% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{check_columns}
\alias{check_columns}
\title{Check all column names are as expected}
\usage{
check_columns(df)
}
\arguments{
\item{df}{data frame output of \link[ddvR]{read_file}}
}
\value{
logical
}
\description{
Check all column names are as expected
}
