% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ddv_import}
\alias{ddv_import}
\title{Run whole import sequence}
\usage{
ddv_import(input)
}
\arguments{
\item{input}{absolute or relative path to the .csv file}
}
\value{
a df with 7 columns
}
\description{
Just strings together all the import functions in the correct order.
}
