% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ddv_transform}
\alias{ddv_transform}
\title{Run whole transform sequence}
\usage{
ddv_transform(df)
}
\arguments{
\item{df}{data frame outpuit from \link[ddvR]{ddv_import}}
}
\value{
df with 15 columns
}
\description{
Run whole transform sequence
}
