% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ely_weeks}
\alias{ely_weeks}
\title{Adds column with iso year and week and ELY}
\usage{
ely_weeks(df)
}
\arguments{
\item{df}{data frame output of import functions}
}
\value{
data frame with 2 more columns than before
}
\description{
Adds the ISO year and week in YYYY-WxX format, because teden and leto are
not enough tu unambiguously determine the correct iso week.
Then also adds the "Equivalent last year" or ELY week, which is the one exactly
52 weeks ago. WHich is the simplest way to do y-o-y stuff I believe.
}
