% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{email_log}
\alias{email_log}
\title{Send email with log}
\usage{
email_log(log, recipient = "maja.zaloznik@gov.si")
}
\arguments{
\item{log}{path to log file}

\item{recipient}{email (not checked) address to be sent to as BCC.
I think single email is all that's allowed. Haven't tried more.}
}
\value{
nothing, side effect is the email being sent.
}
\description{
Convenience wrapper for sending the log
}
