% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fix_types}
\alias{fix_types}
\title{Fix data types}
\usage{
fix_types(df)
}
\arguments{
\item{df}{data frame output of \link[ddvR]{read_file}}
}
\value{
data frame with same dimensions as input
}
\description{
The output of \link[ddvR]{read_file} will usually have only character types
because of a faulty header row. This function fixes the column types, which
also introduces NAs by coersion. These are later removed with
the \link[ddvR]{remove_na_rows} function. Also rename ST_DEJAVNOSTI while i'm
at it
}
