% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{get_data_from_gmail}
\alias{get_data_from_gmail}
\title{Get's the new data from gmail}
\usage{
get_data_from_gmail()
}
\value{
nothing - side effect is downloaded .csv files from the FURS email.
}
\description{
Every Wednesday the data is sent via gp to the umar.data.bot gmail account..
This function checks for it and if it finds it, it downloads the data to
the data folder
}
