% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{override_55}
\alias{override_55}
\title{Manually ovverride some wrong hotel data}
\usage{
override_55(df)
}
\arguments{
\item{df}{data frame output of import functions \link[ddvR]{ddv_transform}}
}
\value{
data frame with same dimensions as input
}
\description{
11.-15.12.2019 there was an odd anomally in the data for the code 55.100
This function manually replaces the data in the original table with some
stuff we kinda made up.
}
