% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_file}
\alias{read_file}
\title{Import csv file}
\usage{
read_file(input)
}
\arguments{
\item{input}{absolute or relative path to .csv file}
}
\value{
data frame.
}
\description{
Reads csv input file using read.csv2
}
