% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate rows}
\usage{
remove_duplicates(df)
}
\arguments{
\item{df}{data frame output of \link[ddvR]{fix_types} or data frame output of \link[ddvR]{remove_na_rows}}
}
\value{
data frame with same number of columns as input and possibly fewer rows
}
\description{
This happened with one of the original bulk import files, had a whole week
duplicated, hence this function.
}
