% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{remove_na_rows}
\alias{remove_na_rows}
\title{Remove rows with NAs}
\usage{
remove_na_rows(df)
}
\arguments{
\item{df}{data frame output of \link[ddvR]{fix_types}}
}
\value{
data frame with same number of columns as input and possibly fewer rows
}
\description{
Remove rows with NAs
}
