% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{remove_xrates}
\alias{remove_xrates}
\title{Remove illegal values for STOPNJA}
\usage{
remove_xrates(df)
}
\arguments{
\item{df}{data frame output of \link[ddvR]{fix_types} or data frame output of \link[ddvR]{remove_na_rows}}
}
\value{
data frame with same number of columns as input and possibly fewer rows
}
\description{
Remove illegal values for STOPNJA
}
