% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{skd_2}
\alias{skd_2}
\title{Add 2 digit SKD code}
\usage{
skd_2(df)
}
\arguments{
\item{df}{data frame output of import functions (add link)}
}
\value{
data frame with 1 more column than before
}
\description{
Extracts 2 digit SKD code from 5 digit one
}
