% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{skd_alpha}
\alias{skd_alpha}
\title{Add alpha SKD code}
\usage{
skd_alpha(df)
}
\arguments{
\item{df}{data frame output of import functions \link[ddvR]{skd_2}}
}
\value{
data frame with 1 more column than before
}
\description{
Adds a single character SKD code from the lookup table.
}
