% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{skd_filter}
\alias{skd_filter}
\title{Add filter for some SKD codes to exclude}
\usage{
skd_filter(df)
}
\arguments{
\item{df}{data frame output of import functions \link[ddvR]{skd_2}}
}
\value{
data frame with 1 more column than before
}
\description{
SOmetimes we like to exclude a set of SKD codes, at the moment they are set
to c(35, 36, 52, 61, 64), This filter is set to 0 for those codes
}
