% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{skd_retail}
\alias{skd_retail}
\title{Adds additional SKD categorisation with retail subgroups}
\usage{
skd_retail(df)
}
\arguments{
\item{df}{data frame output of import functions \link[ddvR]{skd_2}}
}
\value{
data frame with 1 more column than before
}
\description{
This is the standard 2/digit classification with three additional subgroups
for the retail sale sector
}
