% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{skd_zero}
\alias{skd_zero}
\title{Change all empty or NA SKDs into zeros}
\usage{
skd_zero(df)
}
\arguments{
\item{df}{data frame output of import functions \link[ddvR]{ddv_import}}
}
\value{
data frame with sam dimensions as input
}
\description{
Change all empty or NA SKDs into zeros
}
