% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sum_duplicates}
\alias{sum_duplicates}
\title{Sum up rows with duplicate key}
\usage{
sum_duplicates(df)
}
\arguments{
\item{df}{data frame after recoding of rates and SKD}
}
\value{
data frame with same number of columns as input and possibly fewer rows
}
\description{
Add the values for the columns where recoding means they have the same
unique key.
}
