% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{update_ddv}
\alias{update_ddv}
\title{Wrapper function for complete ELT pipeline}
\usage{
update_ddv(
  new_file,
  tbl = "test123",
  email = "maja.zaloznik@gov.si",
  path = "O:/Avtomatizacija/furs-surs-soap/data/"
)
}
\arguments{
\item{new_file}{base filename of the appropriate .csv file that is located in
O:/Avtomatizacija/furs-surs-soap/data/}

\item{tbl}{which table in the database to write to. Default is test123, but the real
one is at davcne_blagajne}

\item{email}{One or more emails to send logs to - as character vector. If NA,
no emails are sent.}

\item{path}{- path for file which might require UNC style}
}
\value{
Nothing, just side effects :). Writes to the database and emails logs.
}
\description{
This wrapper function runs the whole pipeline in the \link[ddvR]{ddv_import} and then
\link[ddvR]{ddv_transform} functions and finally \link[ddvR]{write_to_db}, while
logging everything to the sink and emailing the logs to the listed recipients
}
