% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_to_db}
\alias{write_to_db}
\title{Write data frame to postgres}
\usage{
write_to_db(df, db)
}
\arguments{
\item{df}{dataframe output of \link[ddvR]{ddv_transform}}

\item{db}{name of database on the local server with the davcni_racuni table}
}
\value{
data frame with number of rows in database table.
using mock db for testing.
}
\description{
This code only works with relevant credentials.
}
