% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{date_split}
\alias{date_split}
\title{Add columns that split date into constituent parts}
\usage{
date_split(df)
}
\arguments{
\item{df}{data frame output of import functions (add link)}
}
\value{
data frame with 4 more columns than before
}
\description{
Splits the date into day, month and year columns as well as the ISO
week and adds all four columns
}
