% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ddv_transform}
\alias{ddv_transform}
\title{Run whole transform sequence
Run thourgh all the transformation steps and at the end also change the
colum names to lowercase, because that's how postgres likes them.}
\usage{
ddv_transform(df)
}
\arguments{
\item{df}{data frame outpu from \link[ddvR]{ddv_import}}
}
\value{
df with 20 columns
}
\description{
Run whole transform sequence
Run thourgh all the transformation steps and at the end also change the
colum names to lowercase, because that's how postgres likes them.
}
