% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{remove_xskd}
\alias{remove_xskd}
\title{Remove illegal values for ST_DEJAVNOSTI}
\usage{
remove_xskd(df)
}
\arguments{
\item{df}{data frame output of \link[ddvR]{fix_types} or data frame output of \link[ddvR]{remove_na_rows}}
}
\value{
data frame with same number of columns as input and possibly fewer rows
}
\description{
First remove all rows where the format is not xx.xxx or empty. of the remaining ones
leave the ones in the code table, the rest make empty, since they are unknown
in the new classification.
}
